+++ inst/tinytest/safer/test-atomic_conversions.R
@@ -9,6 +9,7 @@ attr(x, "test") <- "test"
 attr(x, "class") <- "complex"
 myattr <- attributes(x)
 myattr[["class"]] <- NULL
+
 out <- as.logical(x)
 attributes(out) <- myattr
 expect_equal(as_bool(x), out)
@@ -25,6 +26,20 @@ out <- as.character(x)
 attributes(out) <- myattr
 expect_equal(as_chr(x), out)
 
+x <- 1:10
+attr(x, "test") <- "test"
+attr(x, "class") <- "complex"
+myattr <- attributes(x)
+myattr[["class"]] <- NULL
+
+out <- as.complex(x)
+attributes(out) <- myattr
+expect_equal(as_cplx(x), out)
+
+out <- as.raw(x)
+attributes(out) <- myattr
+expect_equal(as_raw(x), out)
+
 
 # test partial vector type casting ====
 x <- seq(-2.5, 2, by =0.5)
@@ -32,6 +47,7 @@ attr(x, "test") <- "test"
 attr(x, "class") <- "complex"
 myattr <- attributes(x)
 myattr[["class"]] <- NULL
+
 out <- expect <- x
 out[1] <- as_bool(out[1])
 expect[1] <- as.logical(out[1])
@@ -52,14 +68,28 @@ out[1] <- as_chr(out[1])
 expect[1] <- as.character(out[1])
 expect_equal(out, expect)
 
+x <- 1:10
+attr(x, "test") <- "test"
+attr(x, "class") <- "complex"
+myattr <- attributes(x)
+myattr[["class"]] <- NULL
+
+out <- expect <- x
+out[1] <- as_cplx(out[1])
+expect[1] <- as.complex(out[1])
+expect_equal(out, expect)
+
+# not doing raw here, as subassignment to/from raw is not supported.
+
+
 
 # test vector chained type casting for all combos ====
 initials <- list(
   sample(c(TRUE, FALSE, size=10, replace=TRUE)),
   1L:10L, seq(-2.5, 2, by =0.5), as.character(1:10)
 )
-as.funs1 <- as.funs2 <- list(as.logical, as.integer, as.double, as.character)
-tiny.funs1 <- tiny.funs2 <- list(as_bool, as_int, as_dbl, as_chr)
+as.funs1 <- as.funs2 <- list(as.logical, as.integer, as.double, as.character, as.complex, as.raw)
+tiny.funs1 <- tiny.funs2 <- list(as_bool, as_int, as_dbl, as_chr, as_cplx, as_raw)
 outs <- list()
 expects<- list()
 iter <- 1
@@ -90,6 +120,7 @@ attr(x, "test") <- "test"
 attr(x, "class") <- "complex"
 myattr <- attributes(x)
 myattr[["class"]] <- NULL
+
 out <- as.logical(x)
 attributes(out) <- myattr
 expect_equal(as_bool(x), out)
@@ -106,6 +137,21 @@ out <- as.character(x)
 attributes(out) <- myattr
 expect_equal(as_chr(x), out)
 
+x <- matrix(1:10, ncol=2)
+colnames(x) <- c("one", "two")
+attr(x, "test") <- "test"
+attr(x, "class") <- "complex"
+myattr <- attributes(x)
+myattr[["class"]] <- NULL
+
+out <- as.complex(x)
+attributes(out) <- myattr
+expect_equal(as_cplx(x), out)
+
+out <- as.raw(x)
+attributes(out) <- myattr
+expect_equal(as_raw(x), out)
+
 
 # test partial matrix type casting ====
 x <- seq(-2.5, 2, by =0.5) |> matrix(ncol=2)
@@ -114,6 +160,7 @@ attr(x, "test") <- "test"
 attr(x, "class") <- "complex"
 myattr <- attributes(x)
 myattr[["class"]] <- NULL
+
 out <- expect <- x
 out[, 1] <- as_bool(out[, 1])
 expect[, 1] <- as.logical(out[, 1])
@@ -134,14 +181,26 @@ out[, 1] <- as_chr(out[, 1])
 expect[, 1] <- as.character(out[, 1])
 expect_equal(out, expect)
 
+x <- matrix(1:10, ncol=2)
+colnames(x) <- c("one", "two")
+attr(x, "test") <- "test"
+attr(x, "class") <- "complex"
+myattr <- attributes(x)
+myattr[["class"]] <- NULL
+out <- expect <- x
+
+out[, 1] <- as_cplx(out[, 1])
+expect[, 1] <- as.complex(out[, 1])
+expect_equal(out, expect)
+
 
 # test matrix chained type casting for all combos ====
 initials <- list(
   sample(c(TRUE, FALSE, size=10, replace=TRUE)),
   1L:10L, seq(-2.5, 2, by =0.5), as.character(1:10)
 )
-as.funs1 <- as.funs2 <- list(as.logical, as.integer, as.double, as.character)
-tiny.funs1 <- tiny.funs2 <- list(as_bool, as_int, as_dbl, as_chr)
+as.funs1 <- as.funs2 <- list(as.logical, as.integer, as.double, as.character, as.complex, as.raw)
+tiny.funs1 <- tiny.funs2 <- list(as_bool, as_int, as_dbl, as_chr, as_cplx, as_raw)
 outs <- list()
 expects<- list()
 iter <- 1