+++ NEWS.md
@@ -2,6 +2,7 @@
 * Added the options "first" and "last" to argument `i` of `strfind()<-`.
 * Tweaked the error messages of `strfind()<-` to be a bit more informative.
 * Slight speed and memory improvement for `stri_locate_ith()`. Also relaxed the length restriction on argument `i` a little bit.
+* Added `as_cplx()` and `as_raw()` to the atomic typecasting functions.
 * Improved the documentation here and there.
 * Tweaked description text.
 * Added more tests.